+++ src/submissions/backend/scripts/__init__.py
@@ -1,0 +1,7 @@
+from .irida import import_irida
+
+def hello(ctx):
+    print("\n\nHello!\n\n")
+
+def goodbye(ctx):
+    print("\n\nGoodbye\n\n")